+++ DREAM/tasdreamExternalTests.hpp
@@ -84,7 +84,7 @@ private:
     int num_mc;
     int wfirst, wsecond, wthird;
     bool verbose;
-    int rngseed;
+    long int rngseed;
 
     TasDREAM::CppUniformSampler u;
